requirements.txt
PyQt5==5.12.2
PyQt5-sip==12.11.0
pyserial==3.5
pywin32==304
